+++ DataLibrary/Models/UserModel.cs
@@ -9,8 +9,8 @@ namespace DataLibrary.Models
         public string Id { get; set; }
         public string UserName { get; set; }
         public string Email { get; set; }
-        public int Height { get; set; }
-        public int Weight { get; set; }
+        public Decimal Height { get; set; }
+        public Decimal Weight { get; set; }
         public int Age { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
